spring10 PostService 에 aop 적용함

+++ Spring10_Example _copy1/src/main/java/com/example/spring10/service/PostServiceImpl.java
@@ -12,6 +12,7 @@ import com.example.spring10.dto.CommentDto;
 import com.example.spring10.dto.CommentListRequest;
 import com.example.spring10.dto.PostDto;
 import com.example.spring10.dto.PostListDto;
+import com.example.spring10.exception.DeniedException;
 import com.example.spring10.repository.CommentDao;
 import com.example.spring10.repository.PostDao;
 
@@ -114,6 +115,16 @@ public class PostServiceImpl implements PostService{
 
 	@Override
 	public void deletePost(long num) {
+		/*
+		//글 작성자와 로그인된 userName 과 다르면 Exception 을 발생시키고 ExceptionController 에서 처리하게 한다.
+		String writer= postDao.getData(num).getWriter();
+		String userName=SecurityContextHolder.getContext().getAuthentication().getName();
+		if(!writer.equals(userName)) {
+			// 상세하게 처리시 ExceptionController 에서 작업 아니면 스프링이 알아서 처리함
+			throw new DeniedException("요청이 거부 되었습니다.");
+		}
+		*/
+		
 		//posts 의 num을 참조하고 있는 자식 레코드를 미리 삭제
 		postDao.deleteReaded(num);
 		//글 삭제
